Create Splitter interface.
