+++ fe/fe-core/src/main/java/org/apache/doris/planner/external/FileSplitStrategy.java
@@ -18,8 +18,7 @@
 package org.apache.doris.planner.external;
 
 import org.apache.doris.common.Config;
-
-import org.apache.hadoop.mapred.FileSplit;
+import org.apache.doris.planner.Split;
 
 public class FileSplitStrategy {
     private long totalSplitSize;
@@ -30,7 +29,7 @@ public class FileSplitStrategy {
         this.splitNum = 0;
     }
 
-    public void update(FileSplit split) {
+    public void update(Split split) {
         totalSplitSize += split.getLength();
         splitNum++;
     }
