+++ fe/fe-core/src/main/java/org/apache/doris/planner/external/LoadScanProvider.java
@@ -18,7 +18,6 @@
 package org.apache.doris.planner.external;
 
 import org.apache.doris.analysis.Analyzer;
-import org.apache.doris.analysis.Expr;
 import org.apache.doris.analysis.ImportColumnDesc;
 import org.apache.doris.analysis.IntLiteral;
 import org.apache.doris.analysis.SlotRef;
@@ -51,9 +50,7 @@ import org.apache.doris.thrift.TScanRangeLocations;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Maps;
-import org.apache.hadoop.mapred.InputSplit;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.Map;
 
@@ -74,11 +71,6 @@ public class LoadScanProvider implements FileScanProviderIf {
 
     @Override
     public TFileType getLocationType() throws DdlException, MetaNotFoundException {
-        return null;
-    }
-
-    @Override
-    public List<InputSplit> getSplits(List<Expr> exprs) throws IOException, UserException {
         return null;
     }
