+++ fe/fe-core/src/main/java/org/apache/doris/planner/external/TVFScanProvider.java
@@ -18,7 +18,6 @@
 package org.apache.doris.planner.external;
 
 import org.apache.doris.analysis.Analyzer;
-import org.apache.doris.analysis.Expr;
 import org.apache.doris.analysis.SlotDescriptor;
 import org.apache.doris.analysis.TupleDescriptor;
 import org.apache.doris.catalog.Column;
@@ -30,7 +29,6 @@ import org.apache.doris.common.UserException;
 import org.apache.doris.load.BrokerFileGroup;
 import org.apache.doris.planner.external.ExternalFileScanNode.ParamCreateContext;
 import org.apache.doris.tablefunction.ExternalFileTableValuedFunction;
-import org.apache.doris.thrift.TBrokerFileStatus;
 import org.apache.doris.thrift.TFileAttributes;
 import org.apache.doris.thrift.TFileFormatType;
 import org.apache.doris.thrift.TFileScanRangeParams;
@@ -38,11 +36,7 @@ import org.apache.doris.thrift.TFileScanSlotInfo;
 import org.apache.doris.thrift.TFileType;
 
 import com.google.common.collect.Lists;
-import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.mapred.FileSplit;
-import org.apache.hadoop.mapred.InputSplit;
 
-import java.io.IOException;
 import java.util.List;
 import java.util.Map;
 
@@ -56,6 +50,7 @@ public class TVFScanProvider extends QueryScanProvider {
         this.tvfTable = tvfTable;
         this.desc = desc;
         this.tableValuedFunction = tableValuedFunction;
+        this.splitter = new TVFSplitter(tableValuedFunction);
     }
 
     public String getFsName() {
@@ -78,18 +73,6 @@ public class TVFScanProvider extends QueryScanProvider {
     @Override
     public TFileType getLocationType() throws DdlException, MetaNotFoundException {
         return tableValuedFunction.getTFileType();
-    }
-
-    @Override
-    public List<InputSplit> getSplits(List<Expr> exprs) throws IOException, UserException {
-        List<InputSplit> splits = Lists.newArrayList();
-        List<TBrokerFileStatus> fileStatuses = tableValuedFunction.getFileStatuses();
-        for (TBrokerFileStatus fileStatus : fileStatuses) {
-            Path path = new Path(fileStatus.getPath());
-            FileSplit fileSplit = new FileSplit(path, 0, fileStatus.getSize(), new String[0]);
-            splits.add(fileSplit);
-        }
-        return splits;
     }
 
     @Override
